backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
transformers==4.30.0
torch==2.2.1
numpy==1.21.0 